refactor: finish moving identity/status ownership to neutral leaves
Complete the R5/R4 extraction so audit/packages aren't in the dependency
path for identity/status concepts a future Applications/Packages surface
will lean on:

- checks.refKey now calls resourceid.ResourceKey instead of re-implementing
  the g|Kind|ns|name format inline (resourceid is stdlib-only, so pkg/checks
  stays k8s-free for Hub).
- Drop the unused audit.GroupForBuiltinKind re-export; repoint its one
  internal caller to resourceid; relocate its test to pkg/resourceid (which
  had no coverage of its own).
- Fix comments that still named audit/packages as the owner of ResourceKey /
  transient-condition semantics.

audit.ResourceKey is kept on purpose — it's paired with audit.IndexByResource
(callers key into an audit-built index with it), a justified surface.
